# frontend/openapi.json
{"components":{"schemas":{"AccountCreate":{"additionalProperties":false,"properties":{"cash_balance":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Cash Balance"},"currency":{"$ref":"#/components/schemas/Currency"},"name":{"maxLength":80,"minLength":1,"title":"Name","type":"string"}},"required":["name","currency"],"title":"AccountCreate","type":"object"},"AccountPatch":{"additionalProperties":false,"properties":{"archived":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Archived"},"cash_balance":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Cash Balance"},"name":{"anyOf":[{"maxLength":80,"minLength":1,"type":"string"},{"type":"null"}],"title":"Name"},"version":{"minimum":0.0,"title":"Version","type":"integer"}},"required":["version"],"title":"AccountPatch","type":"object"},"AccountView":{"properties":{"archived_at":{"anyOf":[{"format":"date-time","type":"string"},{"type":"null"}],"title":"Archived At"},"cash_balance":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Cash Balance"},"created_at":{"format":"date-time","title":"Created At","type":"string"},"currency":{"$ref":"#/components/schemas/Currency"},"id":{"title":"Id","type":"string"},"name":{"title":"Name","type":"string"},"updated_at":{"format":"date-time","title":"Updated At","type":"string"},"version":{"title":"Version","type":"integer"}},"required":["id","name","currency","cash_balance","version","created_at","updated_at","archived_at"],"title":"AccountView","type":"object"},"AdapterStatusView":{"properties":{"enabled":{"title":"Enabled","type":"boolean"},"name":{"title":"Name","type":"string"}},"required":["name","enabled"],"title":"AdapterStatusView","type":"object"},"AnalysisMode":{"enum":["full_mcp","bounded_context","disabled"],"title":"AnalysisMode","type":"string"},"AssetType":{"enum":["equity","etf"],"title":"AssetType","type":"string"},"CompatibilityReport":{"properties":{"analysis_mode":{"$ref":"#/components/schemas/AnalysisMode"},"contract_compatible":{"title":"Contract Compatible","type":"boolean"},"deep_analysis_enabled":{"title":"Deep Analysis Enabled","type":"boolean"},"mcp_status":{"$ref":"#/components/schemas/McpStatus"},"missing_capabilities":{"items":{"type":"string"},"title":"Missing Capabilities","type":"array"},"reasons":{"items":{"type":"string"},"title":"Reasons","type":"array"},"state":{"$ref":"#/components/schemas/CompatibilityState"},"supported_versions":{"default":"<0.2.0,>=0.1.11","title":"Supported Versions","type":"string"},"vibe_version":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Vibe Version"}},"required":["state","analysis_mode","contract_compatible","deep_analysis_enabled","mcp_status"],"title":"CompatibilityReport","type":"object"},"CompatibilityState":{"enum":["compatible","degraded","unsupported"],"title":"CompatibilityState","type":"string"},"Currency":{"enum":["CNY","HKD","USD"],"title":"Currency","type":"string"},"CursorPage_AccountView_":{"properties":{"items":{"items":{"$ref":"#/components/schemas/AccountView"},"title":"Items","type":"array"},"next_cursor":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Next Cursor"}},"required":["items","next_cursor"],"title":"CursorPage[AccountView]","type":"object"},"CursorPage_PositionView_":{"properties":{"items":{"items":{"$ref":"#/components/schemas/PositionView"},"title":"Items","type":"array"},"next_cursor":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Next Cursor"}},"required":["items","next_cursor"],"title":"CursorPage[PositionView]","type":"object"},"ErrorDetail":{"properties":{"code":{"title":"Code","type":"string"},"fields":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Fields"}},"required":["code"],"title":"ErrorDetail","type":"object"},"ErrorEnvelope":{"properties":{"error":{"$ref":"#/components/schemas/ErrorDetail"}},"required":["error"],"title":"ErrorEnvelope","type":"object"},"FailedProbe":{"properties":{"reason":{"default":"probe_execution_failed","title":"Reason","type":"string"},"status":{"$ref":"#/components/schemas/McpStatus","default":"failed"}},"title":"FailedProbe","type":"object"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"title":"Detail","type":"array"}},"title":"HTTPValidationError","type":"object"},"InstrumentConfirm":{"additionalProperties":false,"properties":{"candidate_id":{"format":"uuid","title":"Candidate Id","type":"string"}},"required":["candidate_id"],"title":"InstrumentConfirm","type":"object"},"InstrumentSearchView":{"properties":{"asset_type":{"$ref":"#/components/schemas/AssetType"},"candidate_id":{"format":"uuid","title":"Candidate Id","type":"string"},"canonical_symbol":{"title":"Canonical Symbol","type":"string"},"currency":{"$ref":"#/components/schemas/Currency"},"market":{"$ref":"#/components/schemas/Market"},"name":{"title":"Name","type":"string"},"sources":{"items":{"type":"string"},"title":"Sources","type":"array"}},"required":["candidate_id","canonical_symbol","name","market","currency","asset_type","sources"],"title":"InstrumentSearchView","type":"object"},"InstrumentView":{"properties":{"asset_type":{"$ref":"#/components/schemas/AssetType"},"canonical_symbol":{"title":"Canonical Symbol","type":"string"},"created_at":{"format":"date-time","title":"Created At","type":"string"},"currency":{"$ref":"#/components/schemas/Currency"},"id":{"title":"Id","type":"string"},"market":{"$ref":"#/components/schemas/Market"},"name":{"title":"Name","type":"string"},"updated_at":{"format":"date-time","title":"Updated At","type":"string"}},"required":["id","canonical_symbol","name","market","currency","asset_type","created_at","updated_at"],"title":"InstrumentView","type":"object"},"Market":{"enum":["CN_SH","CN_SZ","CN_BJ","HK","US"],"title":"Market","type":"string"},"McpStatus":{"enum":["not_checked","available","missing","failed"],"title":"McpStatus","type":"string"},"PortfolioSummary":{"properties":{"account_count":{"title":"Account Count","type":"integer"},"currency":{"$ref":"#/components/schemas/Currency"},"estimated":{"title":"Estimated","type":"boolean"},"known_cash":{"title":"Known Cash","type":"string"},"market_value":{"title":"Market Value","type":"string"},"position_cost":{"title":"Position Cost","type":"string"},"position_count":{"title":"Position Count","type":"integer"},"positions":{"items":{"$ref":"#/components/schemas/SummaryPosition"},"title":"Positions","type":"array"},"stale_count":{"title":"Stale Count","type":"integer"},"total_value":{"title":"Total Value","type":"string"},"unknown_cash_account_count":{"title":"Unknown Cash Account Count","type":"integer"},"unrealized_pnl":{"title":"Unrealized Pnl","type":"string"},"unrealized_pnl_pct":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Unrealized Pnl Pct"},"unvalued_cost":{"title":"Unvalued Cost","type":"string"},"unvalued_count":{"title":"Unvalued Count","type":"integer"},"valued_count":{"title":"Valued Count","type":"integer"},"valued_position_cost":{"title":"Valued Position Cost","type":"string"}},"required":["currency","account_count","position_count","valued_count","stale_count","unvalued_count","market_value","position_cost","valued_position_cost","unvalued_cost","unrealized_pnl","unrealized_pnl_pct","known_cash","unknown_cash_account_count","total_value","estimated","positions"],"title":"PortfolioSummary","type":"object"},"PositionCreate":{"additionalProperties":false,"properties":{"account_id":{"format":"uuid","title":"Account Id","type":"string"},"average_cost":{"title":"Average Cost","type":"string"},"instrument_id":{"format":"uuid","title":"Instrument Id","type":"string"},"note":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Note"},"quantity":{"title":"Quantity","type":"string"}},"required":["account_id","instrument_id","quantity","average_cost"],"title":"PositionCreate","type":"object"},"PositionInstrumentView":{"properties":{"asset_type":{"$ref":"#/components/schemas/AssetType"},"canonical_symbol":{"title":"Canonical Symbol","type":"string"},"currency":{"$ref":"#/components/schemas/Currency"},"market":{"$ref":"#/components/schemas/Market"},"name":{"title":"Name","type":"string"}},"required":["canonical_symbol","name","market","currency","asset_type"],"title":"PositionInstrumentView","type":"object"},"PositionPatch":{"additionalProperties":false,"properties":{"archived":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Archived"},"average_cost":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Average Cost"},"note":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Note"},"quantity":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Quantity"},"version":{"minimum":0.0,"title":"Version","type":"integer"}},"required":["version"],"title":"PositionPatch","type":"object"},"PositionView":{"properties":{"account_id":{"title":"Account Id","type":"string"},"archived_at":{"anyOf":[{"format":"date-time","type":"string"},{"type":"null"}],"title":"Archived At"},"average_cost":{"title":"Average Cost","type":"string"},"created_at":{"format":"date-time","title":"Created At","type":"string"},"id":{"title":"Id","type":"string"},"instrument":{"$ref":"#/components/schemas/PositionInstrumentView"},"instrument_id":{"title":"Instrument Id","type":"string"},"note":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Note"},"quantity":{"title":"Quantity","type":"string"},"updated_at":{"format":"date-time","title":"Updated At","type":"string"},"version":{"title":"Version","type":"integer"}},"required":["id","account_id","instrument_id","instrument","quantity","average_cost","note","version","created_at","updated_at","archived_at"],"title":"PositionView","type":"object"},"ProbeError":{"properties":{"code":{"default":"MCP_PROBE_FAILED","title":"Code","type":"string"},"message":{"default":"Portfolio MCP compatibility probe failed","title":"Message","type":"string"}},"title":"ProbeError","type":"object"},"ProbeFailureResponse":{"properties":{"compatibility":{"$ref":"#/components/schemas/CompatibilityReport"},"error":{"$ref":"#/components/schemas/ProbeError"},"probe":{"$ref":"#/components/schemas/FailedProbe"}},"required":["error","probe","compatibility"],"title":"ProbeFailureResponse","type":"object"},"QuoteState":{"enum":["fresh","stale","unavailable"],"title":"QuoteState","type":"string"},"RefreshItemView":{"properties":{"error_code":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error Code"},"instrument_id":{"format":"uuid","title":"Instrument Id","type":"string"},"outcome":{"enum":["updated","stale","unavailable"],"title":"Outcome","type":"string"},"provider":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Provider"}},"required":["instrument_id","outcome","provider","error_code"],"title":"RefreshItemView","type":"object"},"RefreshRequest":{"additionalProperties":false,"properties":{"instrument_ids":{"anyOf":[{"items":{"format":"uuid","type":"string"},"type":"array"},{"type":"null"}],"title":"Instrument Ids"}},"title":"RefreshRequest","type":"object"},"RefreshRunView":{"properties":{"finished_at":{"anyOf":[{"format":"date-time","type":"string"},{"type":"null"}],"title":"Finished At"},"items":{"items":{"$ref":"#/components/schemas/RefreshItemView"},"title":"Items","type":"array"},"providers_used":{"items":{"type":"string"},"title":"Providers Used","type":"array"},"run_id":{"format":"uuid","title":"Run Id","type":"string"},"stale":{"title":"Stale","type":"integer"},"started_at":{"format":"date-time","title":"Started At","type":"string"},"status":{"enum":["running","succeeded","partial","failed"],"title":"Status","type":"string"},"terminal_error":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Terminal Error"},"unavailable":{"title":"Unavailable","type":"integer"},"updated":{"title":"Updated","type":"integer"}},"required":["run_id","status","updated","stale","unavailable","terminal_error","providers_used","started_at","finished_at","items"],"title":"RefreshRunView","type":"object"},"RefreshStatusView":{"properties":{"finished_at":{"format":"date-time","title":"Finished At","type":"string"},"stale":{"title":"Stale","type":"integer"},"status":{"enum":["succeeded","partial","failed"],"title":"Status","type":"string"},"unavailable":{"title":"Unavailable","type":"integer"},"updated":{"title":"Updated","type":"integer"}},"required":["status","updated","stale","unavailable","finished_at"],"title":"RefreshStatusView","type":"object"},"SettingsStatusView":{"properties":{"adapters":{"items":{"$ref":"#/components/schemas/AdapterStatusView"},"title":"Adapters","type":"array"},"backup_directory":{"title":"Backup Directory","type":"string"},"candidate_cache_count":{"title":"Candidate Cache Count","type":"integer"},"database_path":{"title":"Database Path","type":"string"},"last_refresh":{"anyOf":[{"$ref":"#/components/schemas/RefreshStatusView"},{"type":"null"}]},"last_successful_refresh_at":{"anyOf":[{"format":"date-time","type":"string"},{"type":"null"}],"title":"Last Successful Refresh At"},"latest_backup_at":{"anyOf":[{"format":"date-time","type":"string"},{"type":"null"}],"title":"Latest Backup At"},"latest_quote_count":{"title":"Latest Quote Count","type":"integer"},"migration_healthy":{"title":"Migration Healthy","type":"boolean"},"schema_revision":{"title":"Schema Revision","type":"string"}},"required":["schema_revision","migration_healthy","database_path","backup_directory","latest_backup_at","adapters","last_successful_refresh_at","last_refresh","latest_quote_count","candidate_cache_count"],"title":"SettingsStatusView","type":"object"},"SummaryPosition":{"properties":{"account_id":{"title":"Account Id","type":"string"},"allocation":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Allocation"},"average_cost":{"title":"Average Cost","type":"string"},"instrument_id":{"title":"Instrument Id","type":"string"},"market_value":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Market Value"},"position_cost":{"title":"Position Cost","type":"string"},"position_id":{"title":"Position Id","type":"string"},"quantity":{"title":"Quantity","type":"string"},"quote_as_of":{"anyOf":[{"format":"date-time","type":"string"},{"type":"null"}],"title":"Quote As Of"},"quote_fetched_at":{"anyOf":[{"format":"date-time","type":"string"},{"type":"null"}],"title":"Quote Fetched At"},"quote_price":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Quote Price"},"quote_provider":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Quote Provider"},"quote_state":{"$ref":"#/components/schemas/QuoteState"},"unrealized_pnl":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Unrealized Pnl"},"unrealized_pnl_pct":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Unrealized Pnl Pct"}},"required":["position_id","account_id","instrument_id","quantity","average_cost","position_cost","quote_price","market_value","unrealized_pnl","unrealized_pnl_pct","allocation","quote_state","quote_provider","quote_as_of","quote_fetched_at"],"title":"SummaryPosition","type":"object"},"ValidationError":{"properties":{"ctx":{"title":"Context","type":"object"},"input":{"title":"Input"},"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"title":"Location","type":"array"},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}},"required":["loc","msg","type"],"title":"ValidationError","type":"object"}}},"info":{"title":"Vibe-Trading Portfolio","version":"0.1.0"},"openapi":"3.1.0","paths":{"/api/v1/accounts":{"get":{"operationId":"list_accounts_api_v1_accounts_get","parameters":[{"in":"query","name":"archived","required":false,"schema":{"default":false,"title":"Archived","type":"boolean"}},{"in":"query","name":"cursor","required":false,"schema":{"anyOf":[{"format":"uuid","type":"string"},{"type":"null"}],"title":"Cursor"}},{"in":"query","name":"limit","required":false,"schema":{"default":50,"maximum":100,"minimum":1,"title":"Limit","type":"integer"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CursorPage_AccountView_"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"},"503":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorEnvelope"}}},"description":"Service Unavailable"}},"summary":"List Accounts","tags":["portfolio"]},"post":{"operationId":"create_account_api_v1_accounts_post","parameters":[{"in":"header","name":"Idempotency-Key","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Idempotency-Key"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountCreate"}}},"required":true},"responses":{"201":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountView"}}},"description":"Successful Response"},"409":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorEnvelope"}}},"description":"Conflict"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Create Account","tags":["portfolio"]}},"/api/v1/accounts/{account_id}":{"patch":{"operationId":"update_account_api_v1_accounts__account_id__patch","parameters":[{"in":"path","name":"account_id","required":true,"schema":{"format":"uuid","title":"Account Id","type":"string"}},{"in":"header","name":"Idempotency-Key","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Idempotency-Key"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountPatch"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountView"}}},"description":"Successful Response"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorEnvelope"}}},"description":"Not Found"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Update Account","tags":["portfolio"]}},"/api/v1/instruments/confirm":{"post":{"operationId":"confirm_instrument_api_v1_instruments_confirm_post","parameters":[{"in":"header","name":"Idempotency-Key","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Idempotency-Key"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InstrumentConfirm"}}},"required":true},"responses":{"201":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InstrumentView"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorEnvelope"}}},"description":"Unprocessable Content"}},"summary":"Confirm Instrument","tags":["portfolio"]}},"/api/v1/instruments/search":{"get":{"operationId":"search_instruments_api_v1_instruments_search_get","parameters":[{"in":"query","name":"q","required":true,"schema":{"title":"Q","type":"string"}},{"in":"query","name":"limit","required":false,"schema":{"default":10,"title":"Limit","type":"integer"}}],"responses":{"200":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/InstrumentSearchView"},"title":"Response Search Instruments Api V1 Instruments Search Get","type":"array"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorEnvelope"}}},"description":"Unprocessable Content"},"503":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorEnvelope"}}},"description":"Service Unavailable"}},"summary":"Search Instruments","tags":["market-data"]}},"/api/v1/market-data/refresh":{"post":{"operationId":"refresh_quotes_api_v1_market_data_refresh_post","parameters":[{"in":"header","name":"Idempotency-Key","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Idempotency-Key"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshRequest"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshRunView"}}},"description":"Successful Response"},"409":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorEnvelope"}}},"description":"Conflict"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorEnvelope"}}},"description":"Unprocessable Content"},"502":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorEnvelope"}}},"description":"Bad Gateway"}},"summary":"Refresh Quotes","tags":["market-data"]}},"/api/v1/market-data/refresh/{run_id}":{"get":{"operationId":"get_refresh_run_api_v1_market_data_refresh__run_id__get","parameters":[{"in":"path","name":"run_id","required":true,"schema":{"format":"uuid","title":"Run Id","type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshRunView"}}},"description":"Successful Response"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorEnvelope"}}},"description":"Not Found"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Get Refresh Run","tags":["market-data"]}},"/api/v1/portfolio/summary":{"get":{"operationId":"portfolio_summary_api_v1_portfolio_summary_get","parameters":[{"in":"query","name":"currency","required":true,"schema":{"$ref":"#/components/schemas/Currency"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PortfolioSummary"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Portfolio Summary","tags":["portfolio"]}},"/api/v1/positions":{"get":{"operationId":"list_positions_api_v1_positions_get","parameters":[{"in":"query","name":"archived","required":false,"schema":{"default":false,"title":"Archived","type":"boolean"}},{"in":"query","name":"account_id","required":false,"schema":{"anyOf":[{"format":"uuid","type":"string"},{"type":"null"}],"title":"Account Id"}},{"in":"query","name":"cursor","required":false,"schema":{"anyOf":[{"format":"uuid","type":"string"},{"type":"null"}],"title":"Cursor"}},{"in":"query","name":"limit","required":false,"schema":{"default":50,"maximum":100,"minimum":1,"title":"Limit","type":"integer"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CursorPage_PositionView_"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"List Positions","tags":["portfolio"]},"post":{"operationId":"create_position_api_v1_positions_post","parameters":[{"in":"header","name":"Idempotency-Key","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Idempotency-Key"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PositionCreate"}}},"required":true},"responses":{"201":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PositionView"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Create Position","tags":["portfolio"]}},"/api/v1/positions/{position_id}":{"patch":{"operationId":"update_position_api_v1_positions__position_id__patch","parameters":[{"in":"path","name":"position_id","required":true,"schema":{"format":"uuid","title":"Position Id","type":"string"}},{"in":"header","name":"Idempotency-Key","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Idempotency-Key"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PositionPatch"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PositionView"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Update Position","tags":["portfolio"]}},"/api/v1/settings/status":{"get":{"operationId":"settings_status_api_v1_settings_status_get","responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SettingsStatusView"}}},"description":"Successful Response"},"503":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorEnvelope"}}},"description":"Service Unavailable"}},"summary":"Settings Status","tags":["market-data"]}},"/api/v1/system/compatibility":{"get":{"operationId":"compatibility_api_v1_system_compatibility_get","responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CompatibilityReport"}}},"description":"Successful Response"}},"summary":"Compatibility"}},"/api/v1/system/compatibility/mcp-probe":{"post":{"operationId":"probe_mcp_api_v1_system_compatibility_mcp_probe_post","responses":{"200":{"content":{"application/json":{"schema":{}}},"description":"Successful Response"},"502":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProbeFailureResponse"}}},"description":"Bad Gateway"}},"summary":"Probe Mcp"}},"/api/v1/system/status":{"get":{"operationId":"system_status_api_v1_system_status_get","responses":{"200":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"title":"Response System Status Api V1 System Status Get","type":"object"}}},"description":"Successful Response"}},"summary":"System Status"}}}}